requirements.txt
Django==2.2
django-cors-headers==2.5.2
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
mysqlclient==1.4.2.post1
nltk==3.4
psycopg2==2.7.7
pycosat==0.6.3
pycparser==2.19
sqlparse==0.3.0
urllib3==1.24.2
gunicorn==19.9.0
django-heroku==0.3.1
matplotlib
scipy
seaborn
pandas=0.25.0
